+++ reapy_boost/core/reaper/__init__.py
@@ -49,4 +49,5 @@ __all__ = [
     'ui',
     'defer',
     'at_exit',
+    "ExtState",
 ]